requirements.txt
# Flask
Flask==3.0.0
Flask-CORS==4.0.0

# MongoDB
pymongo==4.6.1

# Vector DB
qdrant-client==1.7.3

# ML & NLP
sentence-transformers==2.2.2
torch==2.2.0

# HTTP Requests
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Utilities
numpy==1.26.4
pandas==2.0.3

# Optional (for development)
pytest==7.4.3
black==23.12.0